Fix compiler warnings
Remove 'system("pause")' statement, which is not supported on Linux
platforms and serves no purpose when the program is executed from the
command line.

Previously, a value type of std::exception was being caught in a catch
block, resulting in polymorphic type slicing. Convert this to a capture
by constant reference.